fixing issue with /ser/ during the shortname replacer
